Week_7: updated the lib file for the common use
