+++ Week-7/exercise_7.2/index.js
@@ -1,44 +1,4 @@
-function SinglyLinkedListNode(_value) {
-  this.value = _value;
-  this.next = null;
-}
-function SinglyLinkedList() {
-  this.head = null
-}
-
-SinglyLinkedList.prototype.create = function (_arr) {
-  this.head = new SinglyLinkedListNode(_arr[0])
-  let _current = this.head
-  for (let i = 1; i < _arr.length; ++i) {
-    _current.next = new SinglyLinkedListNode(_arr[i]);
-    _current = _current.next;
-  }
-}
-
-SinglyLinkedList.prototype.traverse = function () {
-  let _current = this.head;
-  while (_current !== null) {
-    console.log(_current.value);
-    _current = _current.next;
-  }
-}
-
-SinglyLinkedList.prototype.rotate = function (_count) {
-  let _prev = this.head,
-    _previousHead = this.head,
-    _current = this.head;
-  let i = 1;
-  while (_current.next) {
-    if (i == _count + 1) {
-      this.head = _current;
-      _prev.next = null;
-    }
-    _prev = _current;
-    _current = _current.next;
-    i++
-  }
-  _current.next = _previousHead;
-}
+const SinglyLinkedList = require('../lib/_linkedList');
 
 const newSll = [2, 4, 7, 8, 9]
 const newSingleLinked = new SinglyLinkedList()
